requirements.txt
rich>=13.0.0
jinja2>=3.1.0
requests>=2.31.0
python-whois>=0.9.4
shodan>=1.31.0
censys>=2.2.0
